move data store logic to DataStor class
