+++ src/include/Kvdb_Impl.h
@@ -22,6 +22,7 @@
 #include "ReadCache.h"
 #include "WorkQueue_.h"
 #include "MetaStor.h"
+#include "DataStor.h"
 
 using namespace dslab;
 namespace hlkvds {
@@ -68,86 +69,21 @@ private:
     void startThds();
     void stopThds();
 
-    Status insertKey(KVSlice& slice);
-    Status updateMeta(Request *req);
-
-    Status readData(KVSlice& slice, string &data);
-
 private:
     SuperBlockManager* sbMgr_;
     IndexManager* idxMgr_;
     BlockDevice* bdev_;
     SegmentManager* segMgr_;
-    GcManager* gcMgr_;
-    dslab::ReadCache* rdCache_;// readcache, rmd160, slru/lru
+
+    ReadCache* rdCache_;// readcache, rmd160, slru/lru
+
     string fileName_;
 
-    SegForReq *seg_;
-    std::mutex segMtx_;
     Options options_;
 
     MetaStor *metaStor_;
+    SimpleDS_Impl *dataStor_;
 
-    // Request Merge thread
-private:
-    class ReqsMergeWQ : public WorkQueue_<Request> {
-    public:
-        explicit ReqsMergeWQ(KVDS *ds, int thd_num=1) : WorkQueue_<Request>(thd_num), ds_(ds) {}
-
-    protected:
-        void _process(Request* req) override {
-            ds_->ReqMerge(req);
-        }
-    private:
-        KVDS *ds_;
-    };
-    ReqsMergeWQ * reqWQ_;
-    void ReqMerge(Request* req);
-
-    // Seg Write to device thread
-private:
-    class SegmentWriteWQ : public WorkQueue_<SegForReq> {
-    public:
-        explicit SegmentWriteWQ(KVDS *ds, int thd_num=1) : WorkQueue_<SegForReq>(thd_num), ds_(ds) {}
-
-    protected:
-        void _process(SegForReq* seg) override {
-            ds_->SegWrite(seg);
-        }
-    private:
-        KVDS *ds_;
-    };
-    SegmentWriteWQ * segWteWQ_;
-    void SegWrite(SegForReq* seg);
-
-    // Seg Timeout thread
-private:
-    std::thread segTimeoutT_;
-    std::atomic<bool> segTimeoutT_stop_;
-    void SegTimeoutThdEntry();
-
-    // Seg Reaper thread
-private:
-    class SegmentReaperWQ : public WorkQueue_<SegForReq> {
-    public:
-        explicit SegmentReaperWQ(KVDS *ds, int thd_num=1) : WorkQueue_<SegForReq>(thd_num), ds_(ds) {}
-
-    protected:
-        void _process(SegForReq* seg) override {
-            ds_->SegReaper(seg);
-        }
-    private:
-        KVDS *ds_;
-    };
-    SegmentReaperWQ *segRprWQ_;
-    void SegReaper(SegForReq* seg);
-
-    //GC thread
-private:
-    std::thread gcT_;
-    std::atomic<bool> gcT_stop_;
-
-    void GCThdEntry();
 };
 
 } // namespace hlkvds
